Implement Palette Swapping
Add a new Pal, PalDisplay and PalReset functions.

Pal replaces color with another one for all subsequent drawings (it is changing the so-called draw palette).

PalDisplay replaces color with another one for the whole screen at the end of a frame (it is changing the so-called display palette).

PalReset resets all swapped colors for all palettes.