Remove Color and ColorReset functions
In order to draw a shape, print text or set a pixel color, game developer has to run two functions: First he has to set a color using pi.Color() and then run a specific drawing function such as pi.Rectfill, pi.Print or pi.Pset. This is unhandy, especially if the color is used only once (Way too many keystrokes needed).

Better solution will be to have the color as a last parameter in all drawing functions. Pico-8 is using such approach. Developer can specify the color for each operation. But there is a catch - specifying color also updates the global color state. So, the next operation will use the new color. This is unexpected (and actually a violation of single responsibility principle).

The idea behind this commit is to get rid of color global state. This will simplify the design and drawing functions will have more obvious behaviour. In the end developer will be less confused.