Move backend code to separate package
All Ebitengine-specific code will be in ebitengine package. This package will provide a backend for running the game loop.

Separating backend code from pi package has many advantages. The most obvious advantage is the ability to choose a different backend. Ebitengine is great, but developer might want to use a different backend for specific platforms or purposes. For example, he might want to use his own backend with a customized look & feel or might want to use a web-browser backend optimized for producing the smallest binaries possible  (with the help of tinygo for example).