Design Audio System
This design makes it possible to use Audio Worklets in the future.

With this design user will be able to play sound effects or music and modify them live, even when sfx or music is currently playing.

Please note that playing sound effects and music is not implemented in this commit yet.